requirements.txt
openai
langchain
psycopg2
pinecone-client